deleting events now work

+++ src/BLL/EventService.java
@@ -3,6 +3,7 @@ package BLL;
 import BE.Event;
 import DAL.AddEvent;
 import DAL.DeleteEvent;
+import DAL.DeleteUserEvent;
 
 import java.util.ArrayList;
 
@@ -10,15 +11,19 @@ public class EventService {
     private final BLLSingleton single = BLLSingleton.getInstance();
     private final DeleteEvent deleteEvent = new DeleteEvent();
     private final AddEvent addEvent = new AddEvent();
+    private final DeleteUserEvent deleteUserEvent = new DeleteUserEvent();
 
     public void addEvent(Event event){
         single.addEventSingle(addEvent.newEvent(event));
     }
 
     public void deleteEvent(Event event){
-        if(deleteEvent.deleteEvent(event)){
-            System.out.println("test");
-            single.deleteEventSingle(event);
+        if(deleteUserEvent.deleteUserEvent(event)){
+            single.deleteUserEvent(event);
+            if(deleteEvent.deleteEvent(event)){
+                System.out.println("test");
+                single.deleteEventSingle(event);
+            }
         }
     }
